+++ profiles/personal/configuration.nix
@@ -53,7 +53,13 @@
 
   xdg.portal = {
     enable = true;
-    extraPortals = [ pkgs.xdg-desktop-portal-gtk ];
+    extraPortals = [ 
+      pkgs.xdg-desktop-portal-gtk
+    ];
+    config = {
+      common.default = "*";
+      hyprland.defaut = ["hyprland" "gtk"];
+    };
   };
 
   programs.hyprland = {